+++ app/screens/edit_profile/constants.ts
@@ -7,33 +7,6 @@ import {t} from '@i18n';
 
 export const MAX_SIZE = 20 * 1024 * 1024;
 
-export const VALID_MIME_TYPES = [
-    'image/jpeg',
-    'image/jpeg',
-    'image/jpg',
-    'image/jp_',
-    'application/jpg',
-    'application/x-jpg',
-    'image/pjpeg',
-    'image/pipeg',
-    'image/vnd.swiftview-jpeg',
-    'image/x-xbitmap',
-    'image/png',
-    'application/png',
-    'application/x-png',
-    'image/bmp',
-    'image/x-bmp',
-    'image/x-bitmap',
-    'image/x-xbitmap',
-    'image/x-win-bitmap',
-    'image/x-windows-bmp',
-    'image/ms-bmp',
-    'image/x-ms-bmp',
-    'application/bmp',
-    'application/x-bmp',
-    'application/x-win-bitmap',
-] as const;
-
 export const FIELDS: { [id: string]: MessageDescriptor } = {
     firstName: {
         id: t('user.settings.general.firstName'),
